# Conflicts:
#	laravel/app/Http/Controllers/FileController.php
#	laravel/routes/web.php

+++ laravel/app/Http/Controllers/FileController.php
@@ -2,47 +2,90 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\FileRequest;
+use App\Models\File;
 use Illuminate\Http\Request;
-use App\Model\File;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
+
 
 class FileController extends Controller
 {
-    public function store(Request $request)
+    public function store(FileRequest $request)
     {
-        // ファイルがアップロードされているか確認
-        if ($request->hasFile('file')) {
+        // バリデーション
+
             // ファイルを保存し、パスを取得
+            $request->validated();
+
+            // ファイルを保存し、パスを取得
+
             $filePath = $request->file('file')->store('public/image');
 
             // ファイルパスをモデルに保存
             $file = new File;
             $file->file_path = str_replace('public/image/', '', $filePath);
+            $file->user_id = Auth::id();
             $file->save();
 
             // 成功した場合のリダイレクト
             return redirect()->route('file.index')->with('success', 'ファイルが正常にアップロードされました。');
-        } else {
-            // ファイルがアップロードされなかった場合のエラーメッセージを返す
-            return redirect()->back()->with('error', 'ファイルを選択してください。');
-        }
+
     }
 
     public function index()
     {
         // データベースからすべてのファイルを取得してビューに渡す
         $files = File::all();
+
+        return view('user.file', compact('files'));
+    }
+
+    public function selectFiles(Request $request)
+    {
+        $selectedFiles = $request->input('selected_files', []);
+        $request->session()->put('selected_files', $selectedFiles);
+
+        return redirect()->route('file.select')->with('success', 'ファイルが選択されました。');
+    }
+
+    public function statusChange(Request $request)
+    {
+        $selectedFiles = $request->session()->get('selected_files', []);
+
+        if (!empty($selectedFiles)) {
+            File::whereIn('id', $selectedFiles)->update(['status' => 1]);
+        }
+
+        // 選択されたファイルの情報を取得し直す
+        $files = File::whereIn('id', $selectedFiles)->get();
+
         return view('user.file', compact('files'));
     }
 
 
+
+
+
+
+
     public function delete($id)
-    {
-        $file = File::find($id);
-        Storage::delete('storage/image/' . $file->file_path);
+        {
 
-        $file->delete();
 
-        return redirect()->route('file.index')->with('success', '画像を削除しました。');
-    }
-}
+            $file = File::find($id);
+
+            $userId = Auth::id();
+            // ログインユーザーが画像の所有者かどうかを確認します
+            if ($file->user_id !== $userId) {
+                return redirect()->back()->with('error', 'この画像は削除できません。');
+
+            }
+
+            Storage::delete('storage/image/'. $file->file_path);
+
+            $file->delete();
+
+            return redirect()->route('file.index')->with('success', '画像を削除しました。');
+
+    }}
